refactor: optimize import & cleanup code
