+++ src/main/java/subway/controller/StationController.java
@@ -1,14 +1,19 @@
 package subway.controller;
 
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
 import java.net.URI;
 import java.util.List;
 
-import subway.service.StationService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
 import subway.dto.StationRequest;
 import subway.dto.StationResponse;
+import subway.service.StationService;
 
 @RestController
 public class StationController {
