+++ src/main/java/subway/domain/Line.java
@@ -5,13 +5,9 @@ import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 import lombok.AllArgsConstructor;
@@ -41,10 +37,10 @@ public class Line {
 
     public static Line of(LineRequest lineRequest, Station upStation, Station downStation) {
         Line line = Line.builder()
-                   .name(lineRequest.getName())
-                   .color(lineRequest.getColor())
-                   .sections(new Sections())
-                   .build();
+                        .name(lineRequest.getName())
+                        .color(lineRequest.getColor())
+                        .sections(new Sections())
+                        .build();
         line.addSection(new Section(upStation, downStation, lineRequest.getDistance()));
         return line;
     }
