+++ src/main/java/subway/domain/Section.java
@@ -1,7 +1,5 @@
 package subway.domain;
 
-import java.util.List;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
