+++ src/main/java/subway/dto/StationResponse.java
@@ -1,11 +1,9 @@
 package subway.dto;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import subway.domain.Section;
 import subway.domain.Sections;
 import subway.domain.Station;
