+++ src/main/java/subway/service/StationService.java
@@ -1,16 +1,16 @@
 package subway.service;
 
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 import java.util.stream.Collectors;
 
-import subway.repository.StationRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import subway.domain.Station;
 import subway.dto.StationRequest;
 import subway.dto.StationResponse;
 import subway.exception.NoStationException;
+import subway.repository.StationRepository;
 
 @Service
 @Transactional(readOnly = true)
@@ -29,19 +29,13 @@ public class StationService {
 
     public List<StationResponse> findAllStations() {
         return stationRepository.findAll().stream()
-                .map(StationResponse::createStationResponse)
-                .collect(Collectors.toList());
+                                .map(StationResponse::createStationResponse)
+                                .collect(Collectors.toList());
     }
 
     @Transactional
     public void deleteStationById(Long id) {
         stationRepository.deleteById(id);
-    }
-
-    public StationResponse findStationById(Long id) {
-        return stationRepository.findById(id)
-                                .map(station -> new StationResponse(station.getId(), station.getName()))
-                                .orElseThrow(() -> new NoStationException("지하철 역이 없습니다."));
     }
 
     public StationResponse findStationById(Long id) {
